♻️ refactor(sabo-cover-ui): extract common Series1 GPIO driver into template
- create template-based `SaboCoverUISeries1GPIO` class to reduce code duplication
- refactor v0.3 and v0.4 drivers to use the new template with hardware-specific parameters
